Git Actions Whit eslint and prettier
